--- game/css/main.css
@@ -1,4 +1,0 @@
-.section:first-of-type{
-	height:100vh;
-	max-height: 800px;
-}
