+++ game/group_start.html
@@ -7,11 +7,21 @@
     <link rel="icon" type="image/png" href="/img/favicon.png">
     <link href="https://fonts.googleapis.com/css?family=Roboto:400,700" rel="stylesheet">
     <script defer src="https://use.fontawesome.com/releases/v5.0.9/js/all.js" integrity="sha384-8iPTk2s/jMVj81dnzb/iFR2sdA7u06vHJyyLlAd4snFpCl/SnyUjRrbdJsw1pGIl" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="/css/common.css"> 
-    <link rel="stylesheet" href="/game/css/game.css">    
-    <!--include your own css like above--> 
-    <script src="/js/common.js" type="module"></script>    
     
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css" integrity="sha384-9gVQ4dYFwwWSjIDZnLEWnxCjeSWFphJiwGPXr1jddIhOegiu1FwO5qRGvFXOdJZ4" crossorigin="anonymous">
+
+    <link rel="stylesheet" href="/css/common.css"> 
+    <link rel="stylesheet" href="/game/css/game.css">
+    <!--include your own css like above--> 
+
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.0/umd/popper.min.js" integrity="sha384-cs/chFZiN24E4KMATLdqdvsezGxaGsi4hLGOzlXwp5UZB1LY//20VyM2taTB4QvJ" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js" integrity="sha384-uefMccjFJAIv6A+rW+L4AHf99KvxDjWSu1z9VI8SKNVmz4sk7buKt/6v9KI65qnm" crossorigin="anonymous"></script>
+    
+    <script src="/js/common.js" type="module"></script>
+    <script src="/game/js/group.js" type="module"></script>    
+    <script src="/game/js/group_start.js" type="module"></script>    
+
     <!--include your own script as a module like above-->   
 </head>
 <body>
@@ -33,9 +43,42 @@
                             Please estimate meal time:
                     </a>
                 </div>
-                <input class= "input small-button" id="meal_time" placeholder="Meal time..">
-                <a class="button small-button" href="group.html" >Submit</a>
-                <a class="button auth-btn" href="group_invite.html" >Invite Another Friend</a>
+
+
+                <div class="button-container">
+                    <div class="inline-buttons">
+
+                        <div class="dropdown show auth-btn dropdown-toggle-split" id="hour-selector">
+                            <a class="btn btn-secondary dropdown-toggle " href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            Hours
+                            </a>
+                        
+                            <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                                <a class="dropdown-item active">0</a>
+                                <a class="dropdown-item">1</a>
+                                <a class="dropdown-item">2</a>
+                                <a class="dropdown-item">3</a>
+                            </div>
+                        </div>
+                        <div class="dropdown show auth-btn dropdown-toggle-split">
+                            <a class="btn btn-secondary dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            Minutes
+                            </a>
+                        
+                            <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                                <a class="dropdown-item">10</a>
+                                <a class="dropdown-item">20</a>
+                                <a class="dropdown-item">30</a>
+                                <a class="dropdown-item active">40</a>
+                                <a class="dropdown-item">50</a>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                
+                <div class="button small-button" href="group.html" >Submit</div>
+                <div class="button small-button" href="group_invite.html" >Invite Another Friend</div>
+                
             </div>
         </div>
     </div>
