+++ index.html
@@ -27,7 +27,6 @@
             <a class="button auth-btn" href="signup.html">Sign Up</a>
         </div>
         
-        
     </div>
 </body>
 </html>
